feat: 공통된 Nav, Footer 컴포넌트 App.js로 분리
