+++ src/App.js
@@ -1,39 +1,14 @@
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { useState, useEffect } from "react";
-import SharedPage from "./pages/SharedPage";
-import FolderPage from "./pages/FolderPage";
-import { getSampleUser, getSampleFolder } from "./api";
-import "./styles/global.css";
+import { Outlet } from "react-router-dom";
+import Nav from "./components/Nav";
+import Footer from "./components/Footer";
 
-const App = () => {
-  const [user, setUser] = useState({});
-  const [folderObj, setFolderObj] = useState({});
-  const [loadingError, setLoadingError] = useState(null);
-
-  const handleLoad = async () => {
-    try {
-      setLoadingError(null);
-      const userBody = await getSampleUser();
-      const folderBody = await getSampleFolder();
-      setUser(userBody);
-      setFolderObj(folderBody);
-    } catch (error) {
-      setLoadingError(error);
-      return;
-    }
-  };
-
-  useEffect(() => {
-    handleLoad();
-  }, []);
-
+const App = ({ user }) => {
   return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/shared" element={<SharedPage user={user} folderObj={folderObj} loadingError={loadingError} />} />
-        <Route path="/folder" element={<FolderPage />} />
-      </Routes>
-    </BrowserRouter>
+    <>
+      <Nav user={user} />
+      <Outlet />
+      <Footer />
+    </>
   );
 };
