+++ src/index.js
@@ -1,12 +1,12 @@
 import { StrictMode } from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App";
+import Main from "./Main";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <>
     <StrictMode>
-      <App />
+      <Main />
     </StrictMode>
   </>
 );
